Split warehouse and admin configs
