+++ docker-compose.yml
@@ -106,10 +106,11 @@ services:
     build:
       context: .
       dockerfile: Dockerfile.static
-    command: bash -c "npx webpack --mode production --watch --info-verbosity verbose"
+    command: bash -c "npx webpack --mode production --config webpack.warehouse.js --watch --info-verbosity verbose"
     volumes:
       - ./warehouse:/opt/warehouse/src/warehouse:z
-      - ./webpack.config.js:/opt/warehouse/src/webpack.config.js:z
+      - ./webpack.common.js:/opt/warehouse/src/webpack.common.js:z
+      - ./webpack.warehouse.js:/opt/warehouse/src/webpack.warehouse.js:z
       - ./.babelrc:/opt/warehouse/src/.babelrc:z
       - ./.sass-lint.yml:/opt/warehouse/src/.sass-lint.yml:z
       - ./tests/frontend:/opt/warehouse/src/tests/frontend:z
